backend/pyproject.toml
[project]
name = "piano-midi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "basic-pitch[onnx]>=0.4.0",
    "fastapi>=0.121.3",
    "flask>=3.1.2",
    "loguru>=0.7.3",
    "musicai-sdk>=1.0.3",
    "pretty-midi>=0.2.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "scalar-fastapi>=1.4.4",
    "tensorflow>=2.15.0",
    "uvicorn>=0.38.0",
]
